% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opdr_code.R
\docType{data}
\name{opdr_code}
\alias{opdr_code}
\title{Corps Code on Opendart}
\format{
\link[tibble:tibble-package]{tibble}
\describe{
\item{corp_name}{character of Corporation name.}
\item{corp_code}{8 number code(but character type) of Corporation code using on opendart api.}
\item{stock_code}{6 number stock code(but character type) of Corporation.}
\item{modify_date}{Date when data updated.}
}
}
\source{
\url{https://opendart.fss.or.kr/api/corpCode.xml}
}
\usage{
opdr_code
}
\description{
Look up airline names from their carrier codes.
}
\examples{
  opdr_code
}
\keyword{datasets}
