% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendartr-package.R
\docType{package}
\name{opendartr-package}
\alias{opendartr}
\alias{opendartr-package}
\title{opendartr: Opendart API wrapper Package}
\description{
This is package for using the opendart api.
}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
